requirements.txt
asgiref==3.5.2
autopep8==2.0.1
bleach==5.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==3.2.16
django-appconf==1.0.5
django-compressor==4.3.1
django-crispy-forms==1.14.0
django-disqus==0.5
django-etc==1.4.0
django-extensions==3.2.1
django-gravatar2==1.4.4
django-hitcount==1.3.5
django-ranged-response==0.2.0
django-robots==5.0
django-simple-captcha==0.5.17
django-summernote==0.8.20.0
django-taggit==3.1.0
django-utils-six==2.0
django-widget-tweaks==1.4.12
django_debug_toolbar==3.8.1
idna==3.4
mock==5.0.1
Pillow==9.3.0
pycodestyle==2.10.0
python-decouple==3.7
pytz==2022.6
rcssmin==1.1.1
requests==2.28.1
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
webencodings==0.5.1
